Add ewma, meter and api_timer
This patch adds a few data structure for derived and accumulative
statistics that are similiar to the yammer implementation used by the
JMX.

ewma - An exponentially-weighted moving average. calculate an event rate
on a given interval.
meter - Calculate 1m, 5m and 15m ewma an all time avrage and a counter.
api_timer - Combines a histogram with a meter. It also expose a
histogram API so it will be an easy task to replace a histogram with a
timer.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>