An installation script for node_exporter
node_exporter is a utility that export node information via prometheus
API. It takes care of host related metrics such as CPU and memory.

The install script, download the node_exporter binaries, create a link
in /usr/bin.

On ubunt it would create a conf file and start it. On RH like systems it
would use an install service file and would start the service.

The installation should be done using sudo.

After a successful installation, the node_exporter would run as a
service.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>