inet_address: add a function to find an ip address from interface
Sometimes it is easier to specify a network interface and let the system
choose the IP address for you.

There are two approaches using the interface name or using an IP range.

Each of this approach has its pros and cons.

The function supports both.

It uses getifaddrs to get the list of available interfaces and loop over
them until an interface with the given name is found or until the
interface ip is in the given range.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>